Refactor the code, so that the constraints are used with the usual GD method
This means, that the "heavy lifting" operations are now performed in the
ConstraintManager class.
