+++ cashocs/_optimization/shape_optimization/mesh_constraints.py
@@ -17,11 +17,12 @@
 
 """Constraints for the mesh / shape in shape optimization."""
 
+from __future__ import annotations
+
 import abc
 
 import fenics
 import numpy as np
-from petsc4py import PETSc
 from scipy import sparse
 
 from cashocs import _utils
@@ -1354,241 +1355,3 @@ class DihedralAngleConstraint(MeshConstraint):
             minimum_angle = cellwise_minimum_angle
 
         return minimum_angle
-
-
-class ConstraintManager:
-    """This class manages all (selected) mesh quality constraints.
-
-    It is used to treat arbitrary constraints (on the mesh coordinates) in a unified
-    manner.
-    """
-
-    def __init__(
-        self,
-        config: cashocs.io.Config,
-        mesh: fenics.Mesh,
-        boundaries: fenics.MeshFunction,
-        deformation_space: fenics.FunctionSpace,
-    ) -> None:
-        """Initializes the constraint manager.
-
-        Args:
-            config: The configuration of the optimization problem.
-            mesh: The underlying FEM mesh.
-            boundaries: The :py:class`fenics.MeshFunction` used to mark the boundaries
-                of the mesh.
-            deformation_space: A :py:class`fenics.FunctionSpace` of vector CG1 elements
-                for mesh deformations.
-
-        """
-        self.config = config
-        self.mesh = mesh
-        self.boundaries = boundaries
-
-        self.has_constraints = (
-            self.config.getfloat("MeshQualityConstraints", "min_angle") > 0.0
-        ) or (
-            self.config.getfloat(
-                "MeshQualityConstraints", "feasible_angle_reduction_factor"
-            )
-            > 0.0
-        )
-        self.v2d = fenics.vertex_to_dof_map(deformation_space)
-        self.d2v = fenics.dof_to_vertex_map(deformation_space)
-        loc0, loc1 = deformation_space.dofmap().ownership_range()
-        self.d2v_local = self.d2v[: loc1 - loc0]
-
-        fun = fenics.Function(deformation_space)
-        self.local_petsc_size = fun.vector().vec().getSizes()[0]
-        self.comm = self.mesh.mpi_comm()
-
-        self.constraints: list[MeshConstraint] = []
-        self.constraint_tolerance: float = self.config.getfloat(
-            "MeshQualityConstraints", "tol"
-        )
-        self.no_constraints = 0
-
-        if self.has_constraints:
-            self.constraints.append(
-                FixedVertexConstraint(
-                    self.mesh, self.boundaries, self.config, deformation_space
-                )
-            )
-
-            if self.mesh.geometry().dim() == 2:
-                self.constraints.append(
-                    TriangleAngleConstraint(self.mesh, self.config, deformation_space)
-                )
-            elif self.mesh.geometry().dim() == 3:
-                self.constraints.append(
-                    DihedralAngleConstraint(self.mesh, self.config, deformation_space)
-                )
-
-        if self.has_constraints:
-            self.inequality_mask = self._compute_inequality_mask()
-            self.no_constraints = len(self.inequality_mask)
-            necessary_constraints = []
-            for constraint in self.constraints:
-                necessary_constraints.append(constraint.is_necessary)
-            self.necessary_constraints = np.concatenate(necessary_constraints)
-
-    def _compute_inequality_mask(self) -> np.ndarray:
-        """Computes the mask which distinguishes between (in-)equality constaints.
-
-        Returns:
-            A boolean array (a mask) which is `True` when the corresponding constraint
-            is an inequality constraint.
-
-        """
-        mask = []
-        for constraint in self.constraints:
-            if constraint.type == "equality":
-                mask += [False] * constraint.no_constraints
-            elif constraint.type == "inequality":
-                mask += [True] * constraint.no_constraints
-
-        return np.array(mask)
-
-    def evaluate(self, coords_seq: np.ndarray) -> np.ndarray:
-        """Evaluates all (selected) mesh quality constraint functions.
-
-        Args:
-            coords_seq: The flattened list of mesh coordinates.
-
-        Returns:
-            A numpy array with the values of the constraint functions.
-
-        """
-        result = []
-        for constraint in self.constraints:
-            result.append(constraint.evaluate(coords_seq))
-
-        if len(result) > 0:
-            self.function_values: np.ndarray = np.concatenate(result)
-        else:
-            return np.array([])
-
-        return self.function_values
-
-    def evaluate_active(
-        self, coords_seq: np.ndarray, active_idx: np.ndarray
-    ) -> np.ndarray:
-        """Evaluates only the constraints corresponding to the working set.
-
-        Args:
-            coords_seq: The flattened list of mesh coordinates.
-            active_idx: The mask of active indices corresponding to the working set.
-
-        Returns:
-            A numpy array containing only the values of the constraint functions that
-            are in the working set.
-
-        """
-        function_values: np.ndarray = self.evaluate(coords_seq)[active_idx]
-        return function_values
-
-    def compute_gradient(self, coords_seq: np.ndarray) -> sparse.csr_matrix:
-        """Computes the gradient of the constraint functions.
-
-        Args:
-            coords_seq: The flattened list of mesh coordinates.
-
-        Returns:
-            A sparse matrix representation of the gradient.
-
-        """
-        result = []
-        for constraint in self.constraints:
-            result.append(constraint.compute_gradient(coords_seq))
-
-        if len(result) > 0:
-            self.gradient = sparse.vstack(result)
-        else:
-            return np.array([])
-
-        return self.gradient
-
-    def compute_active_gradient(
-        self, active_idx: np.ndarray, constraint_gradient: sparse.csr_matrix
-    ) -> PETSc.Mat:
-        """Computes the gradient of those constraint functions that are active.
-
-        Args:
-            active_idx: The boolean array corresponding to the current working set.
-            constraint_gradient: The entire (sparse) constraint gradient.
-
-        Returns:
-            A sparse matrix of the constraint gradient w.r.t. the working set.
-
-        """
-        scipy_matrix = constraint_gradient[active_idx]
-
-        petsc_matrix = _utils.linalg.scipy2petsc(
-            scipy_matrix,
-            self.mesh.mpi_comm(),
-            local_size=self.local_petsc_size,
-        )
-        return petsc_matrix
-
-    def compute_active_set(self, coords_seq: np.ndarray) -> np.ndarray:
-        """Computes the working set.
-
-        The working set is the set of constraints which are active at a given point.
-
-        Args:
-            coords_seq: The list of flattened mesh coordinates.
-
-        Returns:
-            A boolean array which is `True` in component `i` if constraint `i` is
-            active.
-
-        """
-        function_values = self.evaluate(coords_seq)
-        result: np.ndarray = np.abs(function_values) <= self.constraint_tolerance
-        return result
-
-    def is_necessary(self, active_idx: np.ndarray) -> bool:
-        """Checks, whether constraint treatment is necessary based on the working set.
-
-        This can be `False`, e.g., when only fixed boundary constraints are active (as
-        these are directly treated in the gradient deformation) or no constraints are
-        currently active.
-
-        Args:
-            active_idx: The boolean array representing the working set.
-
-        Returns:
-            A boolean, which is `True` when active treatment of the constraints is
-            necessary and `False` otherwise.
-
-        """
-        necessary_constraints_gathered = self.comm.allgather(self.necessary_constraints)
-        necessary_constraints_global = np.concatenate(necessary_constraints_gathered)
-        active_idx_gathered = self.comm.allgather(active_idx)
-        active_idx_global = np.concatenate(active_idx_gathered)
-
-        result = bool(
-            np.any(np.logical_and(necessary_constraints_global, active_idx_global))
-        )
-
-        return result
-
-    def is_feasible(self, coords_seq: np.ndarray) -> np.ndarray:
-        """Checks, whether a given point is feasible w.r.t. all constraints.
-
-        Args:
-            coords_seq: The list of flattened mesh coordinates for the trial point.
-
-        Returns:
-            A boolean array, which is `True` in component `i` if the corresponding
-            constraint is not violated, else `False`.
-
-        """
-        function_values = self.evaluate(coords_seq)
-        function_values_list = self.comm.allgather(function_values)
-        function_values_global = np.concatenate(function_values_list)
-
-        result: np.ndarray = np.less_equal(
-            function_values_global, self.constraint_tolerance
-        )
-        return result
